infer timestamp from entity, move tests, set limit const
