+++ snuba/query/snql/parser.py
@@ -103,6 +103,8 @@ from snuba.query.snql.joins import RelationshipTuple, build_join_clause
 from snuba.state import explain_meta
 from snuba.util import parse_datetime
 
+MAX_LIMIT = 10000
+
 logger = logging.getLogger("snuba.snql.parser")
 
 snql_grammar = Grammar(
@@ -949,7 +951,7 @@ def parse_snql_query_initial(
     limit = parsed.get_limit()
     if limit is None:
         parsed.set_limit(1000)
-    elif limit > 10000:
+    elif limit > MAX_LIMIT:
         raise ParsingException(
             "queries cannot have a limit higher than 10000", should_report=False
         )
